Extract module pyk.prelude.kint
